simplify existing test, and add an even simpler test case
